Rename commands for sbt/mill to generate lsp config
