+++ modules/codegen-plugin/src/sbt-test/codegen-plugin/defaults/test
@@ -14,5 +14,5 @@ $ exists target/scala-2.13/src_managed/main/scala/smithy4s/example/Added.scala
 $ delete src/main/smithy/example.smithy
 -> compile
 
-> generateSmithyBuildJson
+> smithy4sUpdateLSPConfig
 > checkSmithyBuild
